Refactor tor connection
