+++ app/src/main/java/com/tari/android/wallet/tor/TorProxyState.kt
@@ -39,42 +39,10 @@ package com.tari.android.wallet.tor
  */
 sealed class TorProxyState {
     object NotReady : TorProxyState()
+
     object Initializing : TorProxyState()
+
     data class Running(val bootstrapStatus: TorBootstrapStatus) : TorProxyState()
+
     object Failed : TorProxyState()
-}
-
-data class TorBootstrapStatus(val progress: Int, val summary: String, val warning: String? = null) {
-
-    companion object {
-
-        private val warningLogRegex = Regex(".*PROGRESS=(\\d+).*SUMMARY=\"([^\"]*)\".*WARNING=\"([^\"]*)\".*")
-        private val logRegex = Regex(".*PROGRESS=(\\d+).*SUMMARY=\"([^\"]*)\".*")
-        const val maxProgress = 100
-
-        fun from(logLine: String): TorBootstrapStatus {
-            if (warningLogRegex.matches(logLine)) {
-                val matchResult = warningLogRegex.find(logLine)
-                val (progress, summary, warning) = matchResult!!.destructured
-                return TorBootstrapStatus(
-                    progress = progress.toInt(),
-                    summary = summary,
-                    warning = warning
-                )
-            } else if (logRegex.matches(logLine)) {
-                val matchResult = logRegex.find(logLine)
-                val (progress, summary) = matchResult!!.destructured
-                return TorBootstrapStatus(
-                    progress = progress.toInt(),
-                    summary = summary
-                )
-            }
-            return TorBootstrapStatus(
-                progress = 0,
-                summary = "",
-                warning = null
-            )
-        }
-    }
-
 }
